Factored common code between becnhmark.py and test.py.
